prosody-extractor/requirements.txt
# Core dependencies
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
pyyaml==6.0.2

# LLM Provider SDKs
openai==1.55.3
anthropic==0.39.0

# Utilities
requests==2.32.3
tenacity==9.0.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0
httpx==0.27.2  # For FastAPI testing

# Development tools
black==24.10.0
ruff==0.8.0
mypy==1.13.0
